requirements.txt
kafka_python==2.0.2
PyYAML==6.0
